app/src/main/java/com/example/weatherapp/ui/viewmodel/WeatherViewModel.kt
package com.example.weatherapp.ui.viewmodel

import androidx.annotation.VisibleForTesting
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import com.example.weatherapp.data.model.ErrorResponse
import com.example.weatherapp.data.model.WeatherResponse
import com.example.weatherapp.data.repository.WeatherRepository
import dagger.hilt.android.lifecycle.HiltViewModel
import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.StateFlow
import kotlinx.coroutines.launch
import javax.inject.Inject

@HiltViewModel
class WeatherViewModel @Inject constructor(
    private val repository: WeatherRepository
) : ViewModel() {

    private val _weatherState = MutableStateFlow<WeatherResponse?>(null)
    val weatherState: StateFlow<WeatherResponse?> = _weatherState

    private val _errorState = MutableStateFlow<ErrorResponse?>(null)
    val errorState: StateFlow<ErrorResponse?> = _errorState

    fun getWeather(city: String, apiKey: String) {
        viewModelScope.launch {
            repository.getWeather(city, apiKey)
                .onSuccess {
                    _weatherState.value = it
                }
                .onFailure {
                    _errorState.value?.copy(message = "Some thing went wrong, please try again")
                }
        }
    }
}